fix(projects): Soft delete projects when using the delete API
Deletion of projects happened on the queryset. This resulted in a deletion being
executed immediately as part of the request and it could potentially time out
the request.

This fix uses the delete() method on the object. This initially marks the object
with is_deleted=True. After that it schedules a celery task for the deletion
itself.

The test assures that the deletion task is queued and that the soft delete flag
is set correctly.